chatbot-backend/requirements.txt
fastapi
langchain
langchain-community 
langchain-openai
openai
qdrant-client
uvicorn
python-dotenv
playwright
beautifulsoup4
# ...existing dependencies...
